requirements.txt
# NeuroMechFly Simulation Project Dependencies

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Neural simulation (choose one)
# brian2>=2.4.2              # For spiking neural networks
# pytorch>=1.9.0             # For deep learning approaches

# Configuration management
pyyaml>=5.4
hydra-core>=1.1.0           # Optional: for advanced config management

# Physics and visualization
# dm-control>=1.0.0          # For MuJoCo (physics engine)
# neuromechfly>=0.1.0        # NeuroMechFly (once released/available)

# Data I/O and analysis
h5py>=3.0.0                 # For HDF5 data storage
pandas>=1.3.0

# Visualization and analysis
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0              # For interactive notebooks

# Development tools
pytest>=6.2.0               # Testing
black>=21.0                 # Code formatting
flake8>=3.9.0              # Linting
